requirements.txt
boto3==1.37.33
botocore==1.37.33
jmespath==1.0.1
numpy==2.0.2
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
s3transfer==0.11.4
six==1.17.0
tzdata==2025.2
urllib3==1.26.20
